Fix java.util.ConcurrentModificationException
Such exception can happen as the list of locations extracted for each word is sorted locally. If we request a list of locations for the same word in parallel, sorting of the same list happens twice and we get such exception.

To prevent this exception and ensure correct performances, the list of locations is sorted at the end of data files loading.